backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==44.0.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.2.1
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
Pillow==11.1.0
aiofiles==24.1.0
slowapi==0.1.9
phpserialize==1.3
